+++ lang/pt/src/g.ts
@@ -1,9 +1,9 @@
-export const ao: i32 = 2332;
+export const ao: i32 = 2374;
 export const as: i32 = 74;
-export const bm: i32 = 2480;
-export const cm: i32 = 2576;
-export const hv: i32 = 2928;
-export const vm: i32 = 2976;
-export const va: i32 = 3104;
+export const bm: i32 = 2528;
+export const cm: i32 = 2656;
+export const hv: i32 = 3168;
+export const vm: i32 = 3232;
+export const va: i32 = 3408;
 export const lm: i32 = 2;
 export const rm: i32 = 3;