+++ lang/sq/src/g.ts
@@ -1,9 +1,9 @@
-export const ao: i32 = 2274;
+export const ao: i32 = 2310;
 export const as: i32 = 54;
-export const bm: i32 = 2384;
-export const cm: i32 = 2480;
-export const hv: i32 = 2840;
-export const vm: i32 = 2888;
-export const va: i32 = 3016;
+export const bm: i32 = 2424;
+export const cm: i32 = 2520;
+export const hv: i32 = 2880;
+export const vm: i32 = 2928;
+export const va: i32 = 3056;
 export const lm: i32 = 2;
 export const rm: i32 = 2;